requirements.txt
fastapi
uvicorn[standard]
yt-dlp
instagrapi
apscheduler
python-dotenv
